pyproject.toml
[project]
name = "vuln-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "tree-sitter-language-pack>=0.8.0",
]
